Reliability Analysis of Static Structural and  Crash Simulation of Vehicle Architecture.
The objective of this paper is to perform a Reliability Analysis with the primary focus on exit
(output) parameters of Static Structural and Crash analysis simulation (i.e., Yield Stress and Impact Force) of a
Roll-cage vehicle integrated system. For these two types of simulations i.e., Yield Stress Concentration and 
Frontal Crash is performed on a Finite Element (FE) based model with input loading conditions. This was 
performed to observe the Material Behavior in Plastic loading for High strength steel (4130 Chrome Moly) and 
Advance High Strength Steel (Docol r8 Steel). Uncertainty propagation and Quantification was performed to 
find goodness of fit of underlying distribution of a generated random variable using KS and Chi-Square testing. 
Reliability Analysis and Probability of failure of output variable was calculated using FORM and Monte Carlo 
simulations. Sensitivity analysis was performed and Results were benchmarked.